requirements.txt
mmcv_full==1.7.0
mmdet==2.26.0
numpy==1.23.5
pytorch_lightning==1.8.4.post0
torch==1.13.0+cu116
